Move some Errors unit tests out of end-to-end tests
