+++ packages/test/test-end-to-end-tests/src/test/error.spec.ts
@@ -4,7 +4,6 @@
  */
 
 import { strict as assert } from "assert";
-import { ILoggingError } from "@fluidframework/common-definitions";
 import {
     ContainerErrorType,
     LoaderHeader,
@@ -16,15 +15,11 @@ import {
     IFluidResolvedUrl,
     IDocumentServiceFactory,
     DriverErrorType,
-    IThrottlingWarning,
     IUrlResolver,
 } from "@fluidframework/driver-definitions";
-import { createWriteError } from "@fluidframework/driver-utils";
 import {
     createOdspNetworkError,
-    invalidFileNameStatusCode,
 } from "@fluidframework/odsp-doclib-utils";
-import { OdspErrorType } from "@fluidframework/odsp-driver-definitions";
 import { ChildLogger } from "@fluidframework/telemetry-utils";
 import {
     createDocumentId,
@@ -100,20 +95,6 @@ describeNoCompat("Errors Types", (getTestObjectProvider) => {
         }
     });
 
-    //* test CreatecontainerError around errorType precedence, and whether the error is LoggingError or not
-    //* test createOdspNetworkError around the props that should be added
-
-    it("GeneralError Logging Test", async () => {
-        const err = {
-            userData: "My name is Mark",
-            message: "Some message",
-        };
-        const iError = (CreateContainerError(err) as any) as ILoggingError;
-        const props = iError.getTelemetryProperties();
-        assert.equal(props.userData, undefined, "We shouldn't expose the properties of the inner/original error");
-        assert.equal(props.message, err.message, "But name is copied over!");
-    });
-
     function assertCustomPropertySupport(err: any) {
         err.asdf = "asdf";
         if (err.getTelemetryProperties !== undefined) {
@@ -124,177 +105,13 @@ describeNoCompat("Errors Types", (getTestObjectProvider) => {
         }
     }
 
-    it("GenericNetworkError Test_1", async () => {
-        const networkError = createOdspNetworkError("Test Message", 500);
-        assert.equal(networkError.errorType, DriverErrorType.genericNetworkError,
-            "Error should be a genericNetworkError");
-        assertCustomPropertySupport(networkError);
-        assert.equal(networkError.canRetry, true, "default is canRetry");
-    });
-
-    it("GenericNetworkError Test_2", async () => {
-        const networkError = createOdspNetworkError(
-            "Test Message",
-            400 /* statusCode */,
-            undefined /* retryAfterSeconds */);
-        if (networkError.errorType !== DriverErrorType.genericNetworkError) {
-            assert.fail("Error should be a genericNetworkError");
-        }
-        else {
-            assert.equal(networkError.canRetry, false, "400 is non-retryable");
-            assert.equal(networkError.statusCode, 400, "status code should be preserved");
-        }
-    });
-
-    it("GenericNetworkError Test", async () => {
-        const networkError = createOdspNetworkError(
-            "Test Message",
-            500 /* statusCode */);
-        assertCustomPropertySupport(networkError);
-        if (networkError.errorType !== DriverErrorType.genericNetworkError) {
-            assert.fail("Error should be a genericNetworkError");
-        }
-        else {
-            assert.equal(networkError.canRetry, true, "500 is retryable");
-        }
-    });
-
-    it("AuthorizationError Test 401", async () => {
-        const networkError = createOdspNetworkError(
-            "Test Message",
-            401 /* statusCode */);
-        assert.equal(networkError.errorType, DriverErrorType.authorizationError,
-            "Error should be an authorizationError");
-        assertCustomPropertySupport(networkError);
-    });
-
-    it("AuthorizationError Test 403", async () => {
-        const networkError = createOdspNetworkError(
-            "Test Message",
-            403 /* statusCode */);
-        if (networkError.errorType !== DriverErrorType.authorizationError) {
-            assert.fail("Error should be an authorizationError");
-        }
-        else {
-            assert.equal(networkError.errorType, DriverErrorType.authorizationError, "canRetry should be preserved");
-            assert.equal(networkError.canRetry, false, "canRetry should be preserved");
-        }
-    });
-
-    it("OutOfStorageError Test 507", async () => {
-        const networkError = createOdspNetworkError(
-            "Test Message",
-            507 /* statusCode */);
-        assert.equal(networkError.errorType, OdspErrorType.outOfStorageError,
-            "Error should be an OutOfStorageError");
-        assertCustomPropertySupport(networkError);
-    });
-
-    it("FileNotFoundOrAccessDeniedError Test", async () => {
-        const networkError = createOdspNetworkError(
-            "Test Message",
-            404 /* statusCode */);
-        assertCustomPropertySupport(networkError);
-        if (networkError.errorType !== DriverErrorType.fileNotFoundOrAccessDeniedError) {
-            assert.fail("Error should be a fileNotFoundOrAccessDeniedError");
-        }
-        else {
-            assert.equal(networkError.errorType, DriverErrorType.fileNotFoundOrAccessDeniedError,
-                "canRetry should be preserved");
-            assert.equal(networkError.canRetry, false, "canRetry should be preserved");
-        }
-    });
-
-    it("InvalidFileNameError Test 414", async () => {
-        const networkError = createOdspNetworkError(
-            "Test Message",
-            414 /* statusCode */);
-        assert.equal(networkError.errorType, OdspErrorType.invalidFileNameError,
-            "Error should be an InvalidFileNameError");
-        assertCustomPropertySupport(networkError);
-    });
-
-    it("InvalidFileNameError Test", async () => {
-        const networkError = createOdspNetworkError(
-            "Test Message",
-            invalidFileNameStatusCode /* statusCode */);
-        assert.equal(networkError.errorType, OdspErrorType.invalidFileNameError,
-            "Error should be an InvalidFileNameError");
-        assertCustomPropertySupport(networkError);
-    });
-
-    it("ThrottlingError 400 Test", async () => {
-        const networkError = createOdspNetworkError(
-            "Test Message",
-            400 /* statusCode */,
-            100 /* retryAfterSeconds */);
-        assertCustomPropertySupport(networkError);
-        assert.equal(networkError.errorType, DriverErrorType.genericNetworkError, "Error should be a generic");
-        assert.equal((networkError as any).retryAfterSeconds, undefined, "retryAfterSeconds should not be set");
-    });
-
-    it("ThrottlingError Test", async () => {
-        const networkError = createOdspNetworkError(
-            "Test Message",
-            429,
-            100 /* retryAfterSeconds */) as IThrottlingWarning;
-        assertCustomPropertySupport(networkError);
-        assert.equal(networkError.errorType, DriverErrorType.throttlingError, "Error should be a throttlingError");
-        assert.equal(networkError.retryAfterSeconds, 100, "retryAfterSeconds should be preserved");
-    });
-
-    it("WriteError Test", async () => {
-        const writeError = createWriteError("Test Error");
-        assertCustomPropertySupport(writeError);
-        assert.equal(writeError.errorType, DriverErrorType.writeError, "Error should be a writeError");
-        assert.equal(writeError.canRetry, false, "Error should be critical");
-    });
-
-    it("string test", async () => {
-        const text = "Sample text";
-        const writeError = CreateContainerError(text);
-        assertCustomPropertySupport(writeError);
-        assert.equal(writeError.errorType, DriverErrorType.genericError, "Error should be a writeError");
-        assert.equal(writeError.message, text, "Text is preserved");
-    });
-
     it("Check double conversion of network error", async () => {
         const networkError = createOdspNetworkError("Test Error", 400);
         const error1 = CreateContainerError(networkError);
         const error2 = CreateContainerError(error1);
         assertCustomPropertySupport(error1);
         assertCustomPropertySupport(error2);
-        assert.deepEqual(error1, error2, "Both errors should be same!!");
-    });
-
-    it("Check double conversion of general error", async () => {
-        const err = {
-            message: "Test Error",
-        };
-        const error1 = CreateContainerError(err);
-        const error2 = CreateContainerError(error1);
-        assertCustomPropertySupport(error1);
-        assertCustomPropertySupport(error2);
-        assert.deepEqual(error1, error2, "Both errors should be same!!");
-        assert.deepEqual(error2.message, err.message, "Message text should not be lost!!");
-    });
-
-    it("Check frozen error", async () => {
-        const err = {
-            message: "Test Error",
-        };
-        CreateContainerError(Object.freeze(err));
-    });
-
-    it("Preserve existing properties", async () => {
-        const err1 = {
-            errorType: "Something",
-            message: "Test Error",
-            canRetry: true,
-        };
-        const error1 = CreateContainerError(err1);
-        const error2 = CreateContainerError(Object.freeze(error1));
-        assert.equal(error1.errorType, err1.errorType, "Preserve errorType 1");
-        assert.equal(error2.errorType, err1.errorType, "Preserve errorType 2");
+        assert.deepEqual(networkError, error1, "networkError, error1 should be the same!");
+        assert.deepEqual(error1, error2, "error1, error2 should be the same!");
     });
 });
